RM-185: Fix tags and closerequest being able to ping users (#61)
* Add DisableAutoDefer property to command handling

Introduces a DisableAutoDefer property to command registry and updates command execution and HTTP handling logic to support commands that opt out of automatic deferring. This allows commands to respond immediately if possible, improving responsiveness for certain commands.

* Revert 66172b405469b259b04e7943093899a21b17d33a

* Revert 66172b405469b259b04e7943093899a21b17d33a 2

* Handle close request command outside ticket channel

Close request command now sends a message in the ticket channel if executed outside of it, mentioning the ticket user and confirming the request.