RM-211: Add guild owner as admin on create and update (#73)
Add guild owner as admin on create and update

On guild creation, the owner is now added as a ticket admin. Introduced OnGuildUpdate listener to handle ownership changes, adding the new owner as admin and removing admin/support permissions from the previous owner. Registered the new listener in register.go.